Update layout to accept components
